# feat(server): support vectorized warpers in flash causal lm