--- a/server/text_generation_server/models/types.py
+++ b/server/text_generation_server/models/types.py
@@ -21,6 +21,7 @@ class Batch(ABC):
         cls,
         pb: generate_pb2.Batch,
         tokenizer: PreTrainedTokenizerBase,
+        dtype: torch.dtype,
         device: torch.device,
     ) -> "Batch":
         raise NotImplementedError