--- a/server/text_generation_server/server.py
+++ b/server/text_generation_server/server.py
@@ -55,7 +55,7 @@ class TextGenerationService(generate_pb2_grpc.TextGenerationServiceServicer):
 
     async def Prefill(self, request, context):
         batch = self.model.batch_type.from_pb(
-            request.batch, self.model.tokenizer, self.model.device
+            request.batch, self.model.tokenizer, self.model.dtype, self.model.device
         )
 
         generations, next_batch = self.model.generate_token(batch)