--- a/server/text_generation_server/utils/convert.py
+++ b/server/text_generation_server/utils/convert.py
@@ -1,14 +1,10 @@
-import concurrent
-import time
 import datetime
 import torch
 
-from concurrent.futures import ThreadPoolExecutor
 from collections import defaultdict
-from datetime import timedelta
 from loguru import logger
 from pathlib import Path
-from safetensors.torch import load_file, save_file
+from safetensors.torch import save_file
 from safetensors import safe_open
 from typing import Dict, List