@jlamypoirier, adapted your branch to run with flash causal LM so I can try it out in our prod.
Will add you as co-author ofc.