@jlamypoirier 2023-05-17
> It actually seems to be slower at low batch sizes. I will investigate further next week.

Does this have to do with the cuda graph thing? It should be added to the heterogeneous version for a fair comparison (Though I think the graph could include a few more operations?)