@njhill 2023-05-24
@OlivierDehaene what batch size was this using? Would be interesting to see effect on isolated request latencies vs large batch...