@OlivierDehaene 2023-05-25
Ok it's even faster now:

```
✓ queue_time.....................: avg=657.6ms min=1ms     med=651.5ms max=1.43s    p(90)=1.19s    p(95)=1.27s
✓ time_per_token.................: avg=70.62ms min=58ms    med=71ms    max=74ms     p(90)=72ms     p(95)=72ms
  inference_time.................: avg=4.26s   min=3.5s    med=4.28s   max=4.48s    p(90)=4.32s    p(95)=4.37s
```

@njhill, it's faster in all regimes:

| Parameter          | Value              |
|--------------------|--------------------|
| Model              | bigcode/santacoder |
| Sequence Length    | 128                |
| Decode Length      | 60                 |
| N Runs             | 10                 |
| Warmups            | 1                  |
| Temperature        | Some(0.2)          |
| Top K              | None               |
| Top P              | Some(0.95)         |
| Typical P          | None               |
| Repetition Penalty | None               |
| Watermark          | false              |
| Do Sample          | false              |

on main: 

| Step           | Batch Size | Average   | Lowest    | Highest   | p50       | p90       | p99       |
|----------------|------------|-----------|-----------|-----------|-----------|-----------|-----------|
| Prefill        | 1          | 9.84 ms   | 9.67 ms   | 10.35 ms  | 9.81 ms   | 10.35 ms  | 10.35 ms  |
|                | 2          | 16.45 ms  | 16.27 ms  | 16.74 ms  | 16.39 ms  | 16.74 ms  | 16.74 ms  |
|                | 4          | 30.74 ms  | 30.63 ms  | 30.94 ms  | 30.72 ms  | 30.94 ms  | 30.94 ms  |
|                | 8          | 58.24 ms  | 58.06 ms  | 58.59 ms  | 58.30 ms  | 58.59 ms  | 58.59 ms  |
|                | 16         | 108.20 ms | 107.73 ms | 108.89 ms | 108.14 ms | 108.89 ms | 108.89 ms |
|                | 32         | 207.47 ms | 207.09 ms | 208.09 ms | 207.46 ms | 208.09 ms | 208.09 ms |
| Decode (token) | 1          | 8.37 ms   | 8.32 ms   | 8.41 ms   | 8.38 ms   | 8.32 ms   | 8.32 ms   |
|                | 2          | 8.76 ms   | 8.65 ms   | 8.82 ms   | 8.79 ms   | 8.79 ms   | 8.79 ms   |
|                | 4          | 9.19 ms   | 9.16 ms   | 9.22 ms   | 9.18 ms   | 9.20 ms   | 9.20 ms   |
|                | 8          | 10.17 ms  | 10.15 ms  | 10.20 ms  | 10.18 ms  | 10.18 ms  | 10.18 ms  |
|                | 16         | 12.01 ms  | 11.99 ms  | 12.06 ms  | 12.02 ms  | 11.99 ms  | 11.99 ms  |
|                | 32         | 15.86 ms  | 15.78 ms  | 15.89 ms  | 15.86 ms  | 15.89 ms  | 15.89 ms  |
| Decode (total) | 1          | 494.03 ms | 490.93 ms | 496.37 ms | 494.58 ms | 490.93 ms | 490.93 ms |
|                | 2          | 516.59 ms | 510.62 ms | 520.28 ms | 518.84 ms | 518.63 ms | 518.63 ms |
|                | 4          | 542.07 ms | 540.47 ms | 544.03 ms | 541.85 ms | 542.68 ms | 542.68 ms |
|                | 8          | 600.24 ms | 599.12 ms | 601.85 ms | 600.51 ms | 600.75 ms | 600.75 ms |
|                | 16         | 708.86 ms | 707.31 ms | 711.57 ms | 709.31 ms | 707.31 ms | 707.31 ms |
|                | 32         | 935.58 ms | 931.30 ms | 937.57 ms | 935.77 ms | 937.57 ms | 937.57 ms |


| Step    | Batch Size | Average             | Lowest              | Highest             |
|---------|------------|---------------------|---------------------|---------------------|
| Prefill | 1          | 101.68 tokens/secs  | 96.64 tokens/secs   | 103.39 tokens/secs  |
|         | 2          | 121.60 tokens/secs  | 119.46 tokens/secs  | 122.93 tokens/secs  |
|         | 4          | 130.14 tokens/secs  | 129.28 tokens/secs  | 130.58 tokens/secs  |
|         | 8          | 137.37 tokens/secs  | 136.53 tokens/secs  | 137.79 tokens/secs  |
|         | 16         | 147.88 tokens/secs  | 146.93 tokens/secs  | 148.52 tokens/secs  |
|         | 32         | 154.24 tokens/secs  | 153.78 tokens/secs  | 154.52 tokens/secs  |
| Decode  | 1          | 119.43 tokens/secs  | 118.86 tokens/secs  | 120.18 tokens/secs  |
|         | 2          | 228.43 tokens/secs  | 226.80 tokens/secs  | 231.09 tokens/secs  |
|         | 4          | 435.37 tokens/secs  | 433.80 tokens/secs  | 436.66 tokens/secs  |
|         | 8          | 786.35 tokens/secs  | 784.25 tokens/secs  | 787.83 tokens/secs  |
|         | 16         | 1331.71 tokens/secs | 1326.64 tokens/secs | 1334.63 tokens/secs |
|         | 32         | 2018.00 tokens/secs | 2013.71 tokens/secs | 2027.27 tokens/secs |

this branch:

| Step           | Batch Size | Average   | Lowest    | Highest   | p50       | p90       | p99       |
|----------------|------------|-----------|-----------|-----------|-----------|-----------|-----------|
| Prefill        | 1          | 9.98 ms   | 9.78 ms   | 10.79 ms  | 9.87 ms   | 10.79 ms  | 10.79 ms  |
|                | 2          | 16.58 ms  | 16.37 ms  | 17.07 ms  | 16.51 ms  | 17.07 ms  | 17.07 ms  |
|                | 4          | 30.66 ms  | 30.49 ms  | 30.82 ms  | 30.68 ms  | 30.82 ms  | 30.82 ms  |
|                | 8          | 57.67 ms  | 57.42 ms  | 58.07 ms  | 57.58 ms  | 58.07 ms  | 58.07 ms  |
|                | 16         | 106.82 ms | 106.67 ms | 107.07 ms | 106.78 ms | 107.07 ms | 107.07 ms |
|                | 32         | 207.67 ms | 206.81 ms | 208.11 ms | 207.95 ms | 208.11 ms | 208.11 ms |
| Decode (token) | 1          | 8.33 ms   | 8.30 ms   | 8.36 ms   | 8.34 ms   | 8.32 ms   | 8.32 ms   |
|                | 2          | 8.58 ms   | 8.56 ms   | 8.59 ms   | 8.59 ms   | 8.56 ms   | 8.56 ms   |
|                | 4          | 8.81 ms   | 8.78 ms   | 8.84 ms   | 8.83 ms   | 8.82 ms   | 8.82 ms   |
|                | 8          | 9.33 ms   | 9.31 ms   | 9.35 ms   | 9.33 ms   | 9.35 ms   | 9.35 ms   |
|                | 16         | 10.24 ms  | 10.22 ms  | 10.27 ms  | 10.23 ms  | 10.27 ms  | 10.27 ms  |
|                | 32         | 12.59 ms  | 12.55 ms  | 12.63 ms  | 12.59 ms  | 12.63 ms  | 12.63 ms  |
| Decode (total) | 1          | 491.67 ms | 489.55 ms | 493.24 ms | 491.97 ms | 491.17 ms | 491.17 ms |
|                | 2          | 506.14 ms | 505.31 ms | 506.89 ms | 506.54 ms | 505.31 ms | 505.31 ms |
|                | 4          | 520.10 ms | 517.75 ms | 521.72 ms | 520.78 ms | 520.68 ms | 520.68 ms |
|                | 8          | 550.38 ms | 549.32 ms | 551.54 ms | 550.47 ms | 551.54 ms | 551.54 ms |
|                | 16         | 604.00 ms | 602.88 ms | 605.76 ms | 603.82 ms | 605.76 ms | 605.76 ms |
|                | 32         | 742.74 ms | 740.49 ms | 745.39 ms | 742.92 ms | 745.39 ms | 745.39 ms |


| Step    | Batch Size | Average             | Lowest              | Highest             |
|---------|------------|---------------------|---------------------|---------------------|
| Prefill | 1          | 100.24 tokens/secs  | 92.69 tokens/secs   | 102.26 tokens/secs  |
|         | 2          | 120.65 tokens/secs  | 117.19 tokens/secs  | 122.16 tokens/secs  |
|         | 4          | 130.47 tokens/secs  | 129.79 tokens/secs  | 131.17 tokens/secs  |
|         | 8          | 138.73 tokens/secs  | 137.76 tokens/secs  | 139.32 tokens/secs  |
|         | 16         | 149.79 tokens/secs  | 149.43 tokens/secs  | 150.00 tokens/secs  |
|         | 32         | 154.09 tokens/secs  | 153.77 tokens/secs  | 154.73 tokens/secs  |
| Decode  | 1          | 120.00 tokens/secs  | 119.62 tokens/secs  | 120.52 tokens/secs  |
|         | 2          | 233.14 tokens/secs  | 232.79 tokens/secs  | 233.52 tokens/secs  |
|         | 4          | 453.76 tokens/secs  | 452.35 tokens/secs  | 455.81 tokens/secs  |
|         | 8          | 857.58 tokens/secs  | 855.78 tokens/secs  | 859.25 tokens/secs  |
|         | 16         | 1562.92 tokens/secs | 1558.38 tokens/secs | 1565.82 tokens/secs |
|         | 32         | 2541.96 tokens/secs | 2532.90 tokens/secs | 2549.67 tokens/secs |